Move mockData file to data folder
